Proof of concept works with eldritch imports.
