+++ implants/golem/src/inter.rs
@@ -1,4 +1,152 @@
+/*
+ * Copyright 2019 The Starlark in Rust Authors.
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
+// Features we use
+#![feature(box_syntax)]
+#![feature(io_error_more)]
+//
+// Plugins
+#![cfg_attr(feature = "gazebo_lint", feature(plugin))]
+#![cfg_attr(feature = "gazebo_lint", allow(deprecated))] // :(
+#![cfg_attr(feature = "gazebo_lint", plugin(gazebo_lint))]
+// Disagree these are good hints
+#![allow(clippy::type_complexity)]
+
+use std::ffi::OsStr;
+use std::fmt;
+use std::fmt::Display;
+use std::path::PathBuf;
+use std::sync::Arc;
+
+use gazebo::prelude::*;
+use starlark::errors::EvalMessage;
+use starlark::errors::EvalSeverity;
+use starlark::lsp;
+use starlark::read_line::ReadLine;
+use structopt::clap::AppSettings;
+use structopt::StructOpt;
+use anyhow;
+use serde_json;
+
+use eval::Context;
+use eval::ContextMode;
+use types::LintMessage;
+
+mod eval;
+mod types;
+
+#[derive(Debug, StructOpt)]
+#[structopt(
+    name = "starlark",
+    about = "Evaluate Starlark code",
+    global_settings(&[AppSettings::ColoredHelp]),
+)]
+struct Args {
+    #[structopt(
+        long = "lsp",
+        help = "Start an LSP server.",
+        conflicts_with_all = &[
+            "interactive",
+            "dap",
+            "check",
+            "json",
+            "evaluate",
+            "files",
+        ],
+    )]
+    lsp: bool,
+
+    #[structopt(
+        long = "dap",
+        help = "Start a DAP server.",
+        // Conflicts with all options.
+        conflicts_with_all = &[
+            "interactive",
+            "lsp",
+            "check",
+            "json",
+            "extension",
+            "prelude",
+            "evaluate",
+            "files",
+        ],
+    )]
+    dap: bool,
+
+    #[structopt(
+        long = "check",
+        help = "Run checks and lints.",
+        conflicts_with_all = &["lsp", "dap"],
+    )]
+    check: bool,
+
+    #[structopt(
+        long = "json",
+        help = "Show output as JSON lines.",
+        conflicts_with_all = &["lsp", "dap"],
+    )]
+    json: bool,
+
+    #[structopt(
+        long = "extension",
+        help = "File extension when searching directories."
+    )]
+    extension: Option<String>,
+
+    #[structopt(long = "prelude", help = "Files to load in advance.")]
+    prelude: Vec<PathBuf>,
+
+    #[structopt(
+        long = "expression",
+        short = "e",
+        name = "EXPRESSION",
+        help = "Expressions to evaluate.",
+        conflicts_with_all = &["lsp", "dap"],
+    )]
+    evaluate: Vec<String>,
+
+    #[structopt(
+        name = "FILE",
+        help = "Files to evaluate.",
+        conflicts_with_all = &["lsp", "dap"],
+    )]
+    files: Vec<PathBuf>,
+}
+
+// Treat directories as things to recursively walk for .<extension> files,
+// and everything else as normal files.
+// fn expand_dirs(extension: &str, xs: Vec<PathBuf>) -> impl Iterator<Item = PathBuf> {
+//     let extension = Arc::new(extension.to_owned());
+//     xs.into_iter().flat_map(move |x| {
+//         // Have to keep cloning extension so we keep ownership
+//         let extension = extension.dupe();
+//         if x.is_dir() {
+//             Either::Left(
+//                 WalkDir::new(x)
+//                     .into_iter()
+//                     .filter_map(|e| e.ok())
+//                     .filter(move |e| e.path().extension() == Some(OsStr::new(extension.as_str())))
+//                     .map(|e| e.into_path()),
+//             )
+//         } else {
+//             Either::Right(box vec![x].into_iter())
+//         }
+//     })
+// }
 
 #[derive(Default)]
 struct Stats {
@@ -65,265 +213,68 @@ fn interactive(ctx: &Context) -> anyhow::Result<()> {
     }
 }
 
-pub(crate) fn interactive_main() -> anyhow::Result<()> {
+pub fn interactive_main() -> anyhow::Result<()> {
     let mut ctx = Context::new(
-        if args.check {
-            ContextMode::Check
-        } else {
-            ContextMode::Run
-        },
-        !args.evaluate.is_empty() || is_interactive,
-        &expand_dirs(ext, args.prelude).collect::<Vec<_>>(),
-        is_interactive,
+        ContextMode::Run,
+        true,
+        &vec![],
+        true,
     )?;
 
-    Ok(interactive(&ctx)?)
+    interactive(&ctx)?;
+    Ok(())
 }
 
+// fn not_main() -> anyhow::Result<()> {
+//     gazebo::terminate_on_panic();
 
-impl Context {
-    pub(crate) fn new(
-        mode: ContextMode,
-        print_non_none: bool,
-        prelude: &[PathBuf],
-        module: bool,
-    ) -> anyhow::Result<Self> {
-        let globals = globals();
-        let prelude = prelude.try_map(|x| {
-            let env = Module::new();
+//     let args = argfile::expand_args(argfile::parse_fromfile, argfile::PREFIX)?;
+//     let args: Args = Args::from_iter(args);
+//     if args.dap {
+//         dap::server();
+//     } else {
+//         let is_interactive = args.evaluate.is_empty() && args.files.is_empty();
 
-            let mut eval = Evaluator::new(&env);
-            let module = AstModule::parse_file(x, &dialect())?;
-            eval.eval_module(module, &globals)?;
-            env.freeze()
-        })?;
+//         let ext = args
+//             .extension
+//             .as_ref()
+//             .map_or("bzl", |x| x.as_str())
+//             .trim_start_match('.');
+//         let mut ctx = Context::new(
+//             if args.check {
+//                 ContextMode::Check
+//             } else {
+//                 ContextMode::Run
+//             },
+//             !args.evaluate.is_empty() || is_interactive,
+//             &expand_dirs(ext, args.prelude).collect::<Vec<_>>(),
+//             is_interactive,
+//         )?;
 
-        let module = if module {
-            Some(Self::new_module(&prelude))
-        } else {
-            None
-        };
-        let mut builtins: HashMap<LspUrl, Vec<Doc>> = HashMap::new();
-        let mut builtin_symbols: HashMap<String, LspUrl> = HashMap::new();
-        for doc in get_registered_docs() {
-            let uri = Self::url_for_doc(&doc);
-            builtin_symbols.insert(doc.id.name.clone(), uri.clone());
-            builtins.entry(uri).or_default().push(doc);
-        }
-        let builtin_docs = builtins
-            .into_iter()
-            .map(|(u, ds)| (u, render_docs_as_code(&ds).join("\n\n")))
-            .collect();
+//         if args.lsp {
+//             ctx.mode = ContextMode::Check;
+//             lsp::server::stdio_server(ctx)?;
+//         } else if is_interactive {
+//             interactive(&ctx)?;
+//         } else {
+//             let mut stats = Stats::default();
+//             for e in args.evaluate.clone() {
+//                 stats.increment_file();
+//                 drain(ctx.expression(e).messages, args.json, &mut stats);
+//             }
 
-        Ok(Self {
-            mode,
-            print_non_none,
-            prelude,
-            module,
-            builtin_docs,
-            builtin_symbols,
-        })
-    }
+//             for file in expand_dirs(ext, args.files.clone()) {
+//                 stats.increment_file();
+//                 drain(ctx.file(&file).messages, args.json, &mut stats);
+//             }
 
-    fn url_for_doc(doc: &Doc) -> LspUrl {
-        let url = match &doc.item {
-            DocItem::Module(_) => Url::parse("starlark:/native/builtins.bzl").unwrap(),
-            DocItem::Object(_) => {
-                Url::parse(&format!("starlark:/native/builtins/{}.bzl", doc.id.name)).unwrap()
-            }
-            DocItem::Function(_) => Url::parse("starlark:/native/builtins.bzl").unwrap(),
-        };
-        LspUrl::try_from(url).unwrap()
-    }
-
-    fn new_module(prelude: &[FrozenModule]) -> Module {
-        let module = Module::new();
-        for p in prelude {
-            module.import_public_symbols(p);
-        }
-        module
-    }
-
-    fn go(&self, file: &str, ast: AstModule) -> EvalResult<impl Iterator<Item = EvalMessage>> {
-        let mut warnings = Either::Left(iter::empty());
-        let mut errors = Either::Left(iter::empty());
-        let final_ast = match self.mode {
-            ContextMode::Check => {
-                warnings = Either::Right(self.check(&ast));
-                Some(ast)
-            }
-            ContextMode::Run => {
-                errors = Either::Right(self.run(file, ast).messages);
-                None
-            }
-        };
-        EvalResult {
-            messages: warnings.chain(errors),
-            ast: final_ast,
-        }
-    }
-
-    // Convert an anyhow over iterator of EvalMessage, into an iterator of EvalMessage
-    fn err(
-        file: &str,
-        result: anyhow::Result<EvalResult<impl Iterator<Item = EvalMessage>>>,
-    ) -> EvalResult<impl Iterator<Item = EvalMessage>> {
-        match result {
-            Err(e) => EvalResult {
-                messages: Either::Left(iter::once(EvalMessage::from_anyhow(Path::new(file), &e))),
-                ast: None,
-            },
-            Ok(res) => EvalResult {
-                messages: Either::Right(res.messages),
-                ast: res.ast,
-            },
-        }
-    }
-
-    pub(crate) fn expression(
-        &self,
-        content: String,
-    ) -> EvalResult<impl Iterator<Item = EvalMessage>> {
-        let file = "expression";
-        Self::err(
-            file,
-            AstModule::parse(file, content, &dialect()).map(|module| self.go(file, module)),
-        )
-    }
-
-    pub(crate) fn file(&self, file: &Path) -> EvalResult<impl Iterator<Item = EvalMessage>> {
-        let filename = &file.to_string_lossy();
-        Self::err(
-            filename,
-            fs::read_to_string(file)
-                .map(|content| self.file_with_contents(filename, content))
-                .map_err(|e| e.into()),
-        )
-    }
-
-    pub(crate) fn file_with_contents(
-        &self,
-        filename: &str,
-        content: String,
-    ) -> EvalResult<impl Iterator<Item = EvalMessage>> {
-        Self::err(
-            filename,
-            AstModule::parse(filename, content, &dialect()).map(|module| self.go(filename, module)),
-        )
-    }
-
-    fn run(&self, file: &str, ast: AstModule) -> EvalResult<impl Iterator<Item = EvalMessage>> {
-        let new_module;
-        let module = match self.module.as_ref() {
-            Some(module) => module,
-            None => {
-                new_module = Self::new_module(&self.prelude);
-                &new_module
-            }
-        };
-        let mut eval = Evaluator::new(module);
-        eval.enable_terminal_breakpoint_console();
-        let globals = globals();
-        Self::err(
-            file,
-            eval.eval_module(ast, &globals).map(|v| {
-                if self.print_non_none && !v.is_none() {
-                    println!("{}", v);
-                }
-                EvalResult {
-                    messages: iter::empty(),
-                    ast: None,
-                }
-            }),
-        )
-    }
-
-    fn check(&self, module: &AstModule) -> impl Iterator<Item = EvalMessage> {
-        let mut globals = Vec::new();
-        for x in &self.prelude {
-            globals.extend(x.names().map(|s| s.as_str()));
-        }
-        let globals = if self.prelude.is_empty() {
-            None
-        } else {
-            Some(globals.as_slice())
-        };
-
-        module.lint(globals).into_iter().map(EvalMessage::from)
-    }
-}
-
-impl LspContext for Context {
-    fn parse_file_with_contents(&self, uri: &LspUrl, content: String) -> LspEvalResult {
-        match uri {
-            LspUrl::File(uri) => {
-                let EvalResult { messages, ast } =
-                    self.file_with_contents(&uri.to_string_lossy(), content);
-                LspEvalResult {
-                    diagnostics: messages.map(Diagnostic::from).collect(),
-                    ast,
-                }
-            }
-            _ => LspEvalResult::default(),
-        }
-    }
-
-    fn resolve_load(&self, path: &str, current_file: &LspUrl) -> anyhow::Result<LspUrl> {
-        let path = PathBuf::from(path);
-        match current_file {
-            LspUrl::File(current_file_path) => {
-                let current_file_dir = current_file_path.parent();
-                let absolute_path = match (current_file_dir, path.is_absolute()) {
-                    (_, true) => Ok(path),
-                    (Some(current_file_dir), false) => Ok(current_file_dir.join(&path)),
-                    (None, false) => Err(ResolveLoadError::MissingCurrentFilePath(path)),
-                }?;
-                Ok(Url::from_file_path(absolute_path).unwrap().try_into()?)
-            }
-            _ => Err(
-                ResolveLoadError::WrongScheme("file://".to_owned(), current_file.clone()).into(),
-            ),
-        }
-    }
-
-    fn resolve_string_literal(
-        &self,
-        literal: &str,
-        current_file: &LspUrl,
-    ) -> anyhow::Result<Option<StringLiteralResult>> {
-        self.resolve_load(literal, current_file).map(|url| {
-            Some(StringLiteralResult {
-                url,
-                location_finder: None,
-            })
-        })
-    }
-
-    fn get_load_contents(&self, uri: &LspUrl) -> anyhow::Result<Option<String>> {
-        match uri {
-            LspUrl::File(path) => match path.is_absolute() {
-                true => match fs::read_to_string(&path) {
-                    Ok(contents) => Ok(Some(contents)),
-                    Err(e)
-                        if e.kind() == io::ErrorKind::NotFound
-                            || e.kind() == io::ErrorKind::NotADirectory =>
-                    {
-                        Ok(None)
-                    }
-                    Err(e) => Err(e.into()),
-                },
-                false => Err(LoadContentsError::NotAbsolute(uri.clone()).into()),
-            },
-            LspUrl::Starlark(_) => Ok(self.builtin_docs.get(uri).cloned()),
-            _ => Err(LoadContentsError::WrongScheme("file://".to_owned(), uri.clone()).into()),
-        }
-    }
-
-    fn get_url_for_global_symbol(
-        &self,
-        _current_file: &LspUrl,
-        symbol: &str,
-    ) -> anyhow::Result<Option<LspUrl>> {
-        Ok(self.builtin_symbols.get(symbol).cloned())
-    }
-}
+//             if !args.json {
+//                 println!("{}", stats);
+//                 if stats.error > 0 {
+//                     return Err(anyhow::anyhow!("Failed with {} errors", stats.error));
+//                 }
+//             }
+//         }
+//     }
+//     Ok(())
+// }
