+++ implants/golem/src/main.rs
@@ -6,6 +6,7 @@ use tokio::task;
 use std::fs;
 
 use eldritch::{eldritch_run};
+
 mod inter;
 
 async fn run(tome_path: String) -> Result<String, golem::Error> {
